Add an end of support date for XCP-ng 8.1 (#62)
The date of release of XCP-ng 8.3 being still unknown, we can't safely
use it as a moving EOL date for XCP-ng 8.1. Thus support for XCP-ng 8.1
will end at most on 2021-03-31 (which is approximately the date we
initially thought we would release XCP-ng 8.3 at).

This is not a perfect solution, but at least XCP-ng 8.2 LTS is now here
to allow users to benefit from an long term support period.