docs: add multi-cert SB variable instructions
This commit adds instructions for adding multiple certs to a single
variable in the pool-wide variable store for guest secure boot. This
information will be useful for users wishing to use custom certificates
and needing to allocate trust according to a more bespoke trust policy.
It will also be useful to users that would like to add other
certificates in addition to the public MSFT/UEFI certificates.

Signed-off-by: Bobby Eshleman <bobby.eshleman@gmail.com>