Add Mautic integration in order to remove Mailchimp
Our objective is to remove deps on non-EU providers, like Mailchimp.
We prefer to use Open Source solutions and ideally self-hosted.
That's why we'll transition to ListMonk/Mautic to improve our contacts
with our community, while keeping a high level of privacy:
this way, we'll be sure no data will be sell to anyone else.