Added some details on checking IOMMU for PCI passthrough
Add info on how to check if IOMMU is enabled on the current running kernel, and how to add the kernel parameters.
Also add some info related to iommu_groups, and hint on iommu=pt for performance.

I used this to enable PCI passthrough of USB controller, which worked, even if I did not have anything in the /sys/kernel/iommu_groups.

Signed-off-by: Alf Høgemark <alf@i100.no>