Doc: Explain VM disk export from Hyper-V to XCP-ng
A customer provided explanation on how to export a Hyper-V VM disk to
XCP-ng, using PowerShell commands. They also explained in what order
the VM can be converted (before or after exporting). As a result,
two options are now listed:

"If the server can be taken offline" (the VHD can be created from
the VHDX) or "If the server must remain online" (export to VHDX
then convert to VHD).

Source: https://xcp-ng.org/forum/post/86199

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>