storage.md: Highlight storage types with no official support
Some storage types integrations (such as GlusterFS and CephFS) are supported by XCP-ng, but not tested, and Vates does not provide customer support for these. This commit updates the storage type table and the dedicated sections for these SRs to highlight that fact, in order to prevent any potential issue in production environments.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>