Rune to run `koji` in a container
Adding runes to explain how one can run `koji` in a container. The
feature to read a dockerfile from stdin is somewhat recent, but
hopefully old enough for most people. The one-liner is just to be able
to always run `koji`, without needed to prepare a container image
before.

These runes don't allow to load a specfile or other from the current
directory.

Signed-off-by: Anthony PERARD <anthony.perard@vates.tech>