vms.md: reorder the Virtual Machines page
Rather than a separate section dedicated to guest tools, bring guest
tools sections into the parent sections dedicated to the related OS.
Example: "Windows Guest Tools" under "Windows VMs".

Also bring "Windows Guest Tools Security" to the same level as "Windows
Guest Tools" so that it appears in the table of contents on the right.

We keep a small "Guest Tools" section on top of the document to receive
visitors that followed links towards that section.

Some links change. Links previously targeting #windows (not a good
anchor in the context of this page) will not work anymore, but people
will arrive at the top of the page and quickly find the correct link.
The new anchor for Windows Guest Tools is saner: #windows-guest-tools.
Same with Linux, FreeBSD, etc.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>