index.html: stop the slideshow from making the page bounce up/down
For years I've had the index.html page regularly move up and then down,
and I've related this to the slideshow. I think the text (which is not
even displayed in a visible way when the images are present) wraps in
one case on my computer, for some reason (fonts, browser, settings...).

Making it shorter solves it.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>