doc: turn complex procedures into numbered lists
In the XCP-ng technical documentation, at the page called "Create and use custom XCP-NG templates: a guide for Ubuntu", some complex procedures involved multiple steps organized into bullet lists. This can be difficult to follow as we're not quite sure which step we're at. This commit replaces bullet lists with numbered lists so the reader can remember what exact step they're at.

Also, full stops are replaced with semicolons when the step is illustrated by a screenshot or a code sample, for more fluid reading.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>