requirements: Improve vCPU limit
Tested limit is actually the maximum limit. Xen don't allow further than that as
of today. Say "guest may limit" as in practice, the guest will likely boot but with
less vCPUs usable.

Warn about potential performance implications of VMs with a lot of vCPUs and Windows
guests with more than 64 vCPU.

Signed-off-by: Teddy Astie <teddy.astie@vates.tech>