requirements: Clarify NIC limit
Current wording is quite confusing and suggest that operating system has a
inherent NIC limit which is probably not the case (or at least, not at
this limit).
The reality is more likely that lack of netif support means that VM relies
on emulated devices, which has some limitations (i.e we may not expose all
NICs as a emulated device).

Signed-off-by: Teddy Astie <teddy.astie@vates.tech>